RaidDruid Mount and Save
NEW COMMAND! Have a master say "Mount Up" and if you're not inivisible
or paused, the bots who receive the command will summon their mounts.
Useful for when no bard is around, you UseMount is off, and it's faster
to ride rather than run.

/saveini now works with no addition arguments, defaulting to whatever
RDIni is set to. Useful for when you manually /varset something and want
to save the changes without opening notepad.